refactor(editor): extract renderer-agnostic editor into rinch-editor-view (M10 PR1)
The renderer-agnostic editor — RinchDomEditorView, EditorHandle, the Editor
component, the mounted-editor registry, and the caret-blink clock — moves out of
the desktop-gated `rinch` crate into a new wasm-clean crate, `rinch-editor-view`,
so the upcoming web editor (rinch-web) can reuse the SAME model-first view by
projecting it onto its web_sys WebDocument. No behavior change.

The moved code already touched the host only through rinch_core::dom
{DomDocument,NodeHandle,RenderScope} + rinch_editor_core (it is unit-tested against
MockDomDocument), so the move is mechanical. `rinch`'s `editor` module becomes a
re-export (`pub use rinch_editor_view::*`) plus the desktop-only pieces that reach
past the trait into the renderer or a native adapter:

  - block virtualization (Parley estimated_height over rinch-dom's layout tree)
  - AccessKit a11y (the AT-SPI adapter; the derivation itself stays portable)
  - post_remote_delta (the thread-safe collab inbound wrapper; collab_receive_for
    moves to the shared crate)

Registry coupling resolved without a hook: unregister_editor no longer calls
virtualization::forget — pre_layout already prunes windows for editors absent from
all_editors(), and a stale window is never read in between. The handle's
ime_*/hide_overlays methods are promoted to `pub` (the desktop runtime now calls
them across the crate boundary).

Cargo: `collaboration` forwards to rinch-editor-view/collaboration, which holds the
only first-party automerge link — default builds (desktop AND web) still link zero
automerge.

Gates: clippy --workspace --all-targets -D warnings (default + collaboration);
test --workspace (rinch-editor-view 44 default / 54 collab, all green); default
rinch + rinch-editor-view link zero automerge; rinch-editor-view builds clean on
wasm32-unknown-unknown; MCP smoke test — the desktop editor mounts, renders its
model-derived DOM, and types end-to-end.
