+++ Cargo.toml
@@ -30,6 +30,7 @@ rinch-tabler-icons = { path = "crates/rinch-tabler-icons" }
 rinch-dom = { path = "crates/rinch-dom" }
 rinch-editor-core = { path = "crates/rinch-editor-core" }
 rinch-editor-collab = { path = "crates/rinch-editor-collab" }
+rinch-editor-view = { path = "crates/rinch-editor-view" }
 rinch-debug = { path = "crates/rinch-debug" }
 rinch-editable = { path = "crates/rinch-editable" }
 rinch-platform = { path = "crates/rinch-platform" }
